refactor: 경로별로 디렉터리 분리
